combined/requirements.txt
# Combined Pirate Assistant Requirements
# STT Dependencies
vosk==0.3.44
sounddevice==0.5.2

# LLM Dependencies  
openai>=1.99.0

# TTS Dependencies
elevenlabs==1.4.0

# Utility Dependencies
python-dotenv==1.1.0

# Audio Processing
cffi==1.17.1